refactor(shared): centralize trace detail path builder
